Many bug fixes and enhancements commited as Tag revision

+++ TargomanSMT/libsrc/libTargomanSMT/Private/RuleTable/clsMosesPlainRuleTable.h
@@ -69,10 +69,10 @@ private:
     int PrecomputedValueIndex = -1;
 
 private:
-    static Targoman::Common::Configuration::tmplConfigurable<FilePath_t>   PhraseTableFilePath;            /**< File name of phrase table. */
-    static Targoman::Common::Configuration::tmplConfigurable<FilePath_t>   ReorderingTableFilePath;        /**< File name of reordering table. */
-    static Targoman::Common::Configuration::tmplConfigurable<QString>   WordAlignmentFilePath;        /**< File name of word level alignment of phrases. */
-    static Targoman::Common::Configuration::tmplConfigurable<int>       MaxRuleNodeTargetRuleCount;     /**< Maximum number of target rules kept for each rule node. */
+    static Targoman::Common::Configuration::tmplConfigurable<FilePath_t>   PhraseTableFilePath;         /**< File name of phrase table. */
+    static Targoman::Common::Configuration::tmplConfigurable<FilePath_t>   ReorderingTableFilePath;     /**< File name of reordering table. */
+    static Targoman::Common::Configuration::tmplConfigurable<QString>      WordAlignmentFilePath;       /**< File name of word level alignment of phrases. */
+    static Targoman::Common::Configuration::tmplConfigurable<quint16>      MaxRuleNodeTargetRuleCount;  /**< Maximum number of target rules kept for each rule node. */
 
     TARGOMAN_DEFINE_MODULE(MosesPlainRuleTable);
 };
